requirements.txt
Flask>=0.10.1
MarkupSafe>=0.23
mock>=1.3.0
Werkzeug>=0.7
Jinja2>=2.4
itsdangerous>=0.21
requests
pbr>=0.11
six>=1.7
python-dotenv==0.4.0
gunicorn==19.4.5
